Write a C program that prints the size of various types on the computer it is compiled and run on

+++ 0x00-hello_world/6-size.c
@@ -10,7 +10,7 @@ int main(void)
 {
 int intT;
 char charT;
-long longT;
+double longT;
 float floatT;
 
 printf("Size of a char: %2d byte(s)\n", sizeof(charT));
